refactor: 프로젝트 관리 common, guest, host 폴더 정리
